@@ -1,8 +1,17 @@
-﻿namespace Assets.Scripts.PlayerControl
+﻿using System.Collections.Generic;
+using RebuildSharedData.ClientTypes;
+using RebuildSharedData.Enum;
+
+namespace Assets.Scripts.PlayerControl
 {
-    class PlayerState
+    public class PlayerState
     {
+        public bool IsValid = false;
         public int Level;
         public int Exp;
+        public ClientSkillTree SkillTree;
+        public int SkillPoints;
+        public int JobId;
+        public Dictionary<CharacterSkill, int> KnownSkills = new();
     }
 }
