@@ -50,7 +50,7 @@ public class PacketEnterServer : IClientPacketHandler
         networkPlayer.Connection = connection;
         connection.Player = networkPlayer;
 
-        connection.Player.UpdateStats();
+        //connection.Player.UpdateStats();
 
         ServerLogger.Debug($"Player assigned entity {playerEntity}, creating entity at location {connection.Character.Position}.");
 
